backend/requirements.txt
# FastAPI and web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# YAML configuration
pyyaml>=6.0.1
jsonschema>=4.21.0

# AI/ML
openai>=1.10.0
pytesseract>=0.3.10
pdf2image>=1.16.3
pillow>=10.2.0

# Cloud storage
boto3>=1.34.0  # AWS S3
google-cloud-storage>=2.14.0  # GCS (optional)

# Email
sendgrid>=6.11.0

# Templating
jinja2>=3.1.3

# Background tasks / scheduling
celery>=5.3.6
redis>=5.0.1
apscheduler>=3.10.4

# Utilities
python-dateutil>=2.8.2
httpx>=0.26.0

# Testing
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
httpx>=0.26.0

# Development
black>=24.1.0
isort>=5.13.2
mypy>=1.8.0
ruff>=0.1.14
